patient_logger_agent/pyproject.toml
[project]
name = "rotterdam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "google-cloud-datastore>=2.5.0",
]
